pyproject.toml
[project]
name = "2-prompt-engineering"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.7.0",
    "groq>=1.0.0",
    "python-dotenv>=1.2.1",
]
